[master] 1/4 Help page
Description

A version control system is software that can track the changes implemented in a program. There are several popular version control systems (like Git, SVN, or Mercurial). Each of them has its pros and cons. They share one common idea. A version control system remembers who changed the file when it was done, and why. It allows you to roll back to the previous versions as well.

In this project, you need to implement a simple version control system. It will be your own Git. By the way, if you want to learn how to work with Git, take a look at the Git How To.

Take a look at the commands you will need to implement during this project:

1. config sets or outputs the name of a commit author;
2. --help prints the help page;
3. add adds a file to the list of tracked files or outputs this list;
4. log shows all commits;
5. commit saves file changes and the author name;
6. checkout allows you to switch between commits and restore a previous file state.

In this stage, your program should be able to accept a single argument and, depending on the argument, print help information.
Objectives

In this stage, your program should:

1. Take one argument as a command.
2. If an argument is missing or it is --help, print the entire help page.
3. If a command exists, the program should output the description of the command.
4. If the command is wrong, print '[passed argument]' is not a SVCS command.