[master] 2/4 Add & config
Description

In this stage, your program should allow users to set their name and add the files they want to track. Store a username in config.txt.

Index.txt stores the files that were added to the index. Don't forget to store all the files of the version control system in the vcs directory. You should create this directory programmatically.

Objectives

You need to work on the following commands:

1. config should allow the user to set their name or output an already existing name. If a user wants to set a new name, the program must overwrite the old one.
2. add should allow the user to set the name of a file they want to track or output the names of tracked files. If the file does not exist, the program should inform a user that the file does not exist.