[master] 4/4 Checkout time
Description

In the last stage, implement the checkout command. It allows a user to switch between commits and restore the contents of the files according to the current commit.

Objectives

The checkout command must be passed to the program together with the commit ID to indicate which commit should be used. If a commit with the given ID exists, the contents of the tracked file should be restored in accordance with this commit.